ColumnExtract: add compilation unit
